+++ basis/appseed/openssl/ssl/s3_srvr.c
@@ -1174,7 +1174,7 @@ int ssl3_get_client_hello(SSL *s)
 	/* TLS extensions*/
 	if (s->version >= SSL3_VERSION)
 		{
-		if (!ssl_parse_clienthello_tlsext(s,&p,d,n, &al))
+		if (!ssl_parse_clienthello_tlsext(s,&p,d, (int) n, &al))
 			{
 			/* 'al' set by ssl_parse_clienthello_tlsext */
 			SSLerr(SSL_F_SSL3_GET_CLIENT_HELLO,SSL_R_PARSE_TLSEXT);
@@ -1507,8 +1507,8 @@ int ssl3_send_server_hello(SSL *s)
 
 		s->state=SSL3_ST_SW_SRVR_HELLO_B;
 		/* number of bytes to write */
-		s->init_num=p-buf;
-		s->init_off=0;
+		s->init_num = (int) (p - buf);
+		s->init_off = 0;
 		}
 
 	/* SSL3_ST_SW_SRVR_HELLO_B */
@@ -1746,7 +1746,7 @@ int ssl3_send_server_key_exchange(SSL *s)
 			 * First check the size of encoding and
 			 * allocate memory accordingly.
 			 */
-			encodedlen = EC_POINT_point2oct(group, 
+			encodedlen = (int) EC_POINT_point2oct(group,
 			    EC_KEY_get0_public_key(ecdh),
 			    POINT_CONVERSION_UNCOMPRESSED, 
 			    NULL, 0, NULL);
@@ -1761,7 +1761,7 @@ int ssl3_send_server_key_exchange(SSL *s)
 				}
 
 
-			encodedlen = EC_POINT_point2oct(group, 
+			encodedlen = (int) EC_POINT_point2oct(group,
 			    EC_KEY_get0_public_key(ecdh), 
 			    POINT_CONVERSION_UNCOMPRESSED, 
 			    encodedPoint, encodedlen, bn_ctx);
@@ -3464,7 +3464,7 @@ int ssl3_send_newsession_ticket(SSL *s)
 		p += hlen;
 		/* Now write out lengths: p points to end of data written */
 		/* Total length */
-		len = p - (unsigned char *)s->init_buf->data;
+		len = (int) (p - (unsigned char *)s->init_buf->data);
 		p=(unsigned char *)s->init_buf->data + 1;
 		l2n3(len - 4, p); /* Message length */
 		p += 4;
